refactor: session page modularization with hooks and components
